% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_module.R
\name{managerUI}
\alias{managerUI}
\alias{managerServer}
\title{Content manager shiny module}
\usage{
managerUI(id)

managerServer(input, output, session, .cm)
}
\description{
Content manager shiny module
}
